another fix

+++ vm/commands.c
@@ -11,7 +11,7 @@ int		op_live(t_op_context *c, void *arg1, void *arg2, void *arg3)
 	c->proc->last_live = c->vm->i;
 	reg_value = read_uint(c->vm, arg1, REG_SIZE);
 	champ_id = UINT_MAX - reg_value - 1;
-	if (champ_id == c->proc->champ_id && champ_id < c->vm->n_champs)
+	if (champ_id == c->proc->champ_id && champ_id < (uint)c->vm->n_champs)
 		c->vm->winner = champ_id;
 	return (1);
 }
